function-code/requirements.txt
google-cloud-bigquery
requests
pandas
pyarrow 